--- src/main/java/racing_car_with_winner/GameInput.java
@@ -1,8 +1,0 @@
-package racing_car_with_winner;
-
-public class GameInput {
-
-    public static String getName() {
-        return "name";
-    }
-}
