+++ src/main/java/racing_car_with_winner/GameMachine.java
@@ -3,6 +3,5 @@ package racing_car_with_winner;
 public class GameMachine {
 
     public static void main(String[] args) {
-
     }
 }
