+++ src/main/java/racing_car_with_winner/model/Car.java
@@ -1,4 +1,4 @@
-package racing_car_with_winner;
+package racing_car_with_winner.model;
 
 public class Car {
     private CarName carName;
