+++ src/main/java/racing_car_with_winner/model/GameAttemptCount.java
@@ -1,4 +1,4 @@
-package racing_car_with_winner;
+package racing_car_with_winner.model;
 
 
 import racing_car_with_winner.exception.InvalidInputException;
