+++ src/main/java/racing_car_with_winner/view/UserInterface.java
@@ -1,4 +1,4 @@
-package racing_car_with_winner;
+package racing_car_with_winner.view;
 
 public interface UserInterface {
     void show(String message);
