+++ src/test/java/racing_car_with_winner/model/GameAttemptCountTest.java
@@ -1,12 +1,11 @@
-package racing_car_with_winner;
-
-import static org.junit.jupiter.api.Assertions.*;
+package racing_car_with_winner.model;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import racing_car_with_winner.exception.InvalidInputException;
+import racing_car_with_winner.model.GameAttemptCount;
 
 class GameAttemptCountTest {
